API documentation

Organizing and formatting data

LexiconMining.formatentries — Function

Read source file f and format content for mainentries.cex

formatentries(f)

LexiconMining.typelist — Function

Read source file f and determine list of unique values for @type attribute on entries.

typelist(f)

Determine list of unique values for @type attribute on entries in lns, a Vector of source data lines.

typelist(lns, reader)

LexiconMining.pos_itype_counts — Function

Read morphological data file f and compute occurrences of pos*itype.

pos_itype_counts(f)

Compute occurrences of pos*itype in lns, a Vector of morphological data lines.

pos_itype_counts(lns, reader)

Working with extracted morphological data

Missing docstring.

Missing docstring for MorphData. Check Documenter's build log for details.

Missing docstring.

Missing docstring for morphData. Check Documenter's build log for details.